feat: move client.sd to kotlin root
