+++ src/main/kotlin/net/atos/client/sd/exception/BadRequestException.kt
