+++ src/main/kotlin/net/atos/client/sd/exception/BadRequestExceptionMapper.kt
