+++ src/main/kotlin/net/atos/client/sd/model/WizardResponse.kt
